experiments/gradio_exp/multi_utility_app/pyproject.toml
[project]
name = "multi-utility-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gitingest>=0.1.2",
    "gradio>=5.0.0",
    "langchain-openai>=0.3.0",
    "langchain>=0.3.14",
    "openai>=1.59.6",
    "python-dotenv>=1.0.1",
    "pytube>=15.0.0",
    "torch>=2.5.1",
    "transformers>=4.48.0",
    "yt-dlp>=2024.12.23",
]
